Work around katib-controller behavior
The katib-controller code really, really wants to overwrite bits of
Kubernetes infrastructure, which includes bits of infra that the charm
sets up via Juju. The upstream code also is hardcoded to set the service
up as using the label `app: katib-controller`, whereas Juju labels the
pod as `juju-app: katib-controller`.

So, the upstream code got forked to just set up the infrastructure
correctly, and this charm can just use the r.c.c docker image that uses
that forked code.