pin CI aws VM to a type instead of resources
For AWS, if Juju VMs ask for constraints of cpu/memory it is non-deterministic which vm type Juju requests, sometimes resulting in machines that have >3x the cost of a simple machine.  This change pins to a specific instance-type that is suitable for tests and inexpensive.

See also https://bugs.launchpad.net/juju/+bug/1958574