fix: integration tests for non-existent users
Tests previously created a Namespace but not a Profile for the username to use for testing.  This resulted in missing roles (specifically here, the ones needed to run pipelines), thus testing with any user other than `admin` (which has a default Profile created) would fail pipeline tests.  This fix creates a Profile object for whatever user is specified in the `--username` test suite arg