fix: kf_authentication.py fails for some urls
The previous iteration worked if the url passed to `kubeflow_login` was the kubeflow host domain (eg: `http://my.kubeflow/`), but not if the url was any other valid url under the domain (eg: `http://my.kubeflow/pipelines`).  This was because some of the `request.get`s in the authentication flow needed the domain rather than the target url.  This was missed because only the domain was tested when this was first written.

These updates fix this problem - the authentication now works for all valid urls in the domain.  Note that authentication will not work for any page that does not exist (eg: `http://my.kubeflow/somethingThatDoesNotExist` will **not** result in authentication)