feat: scan of images in bundle (#634)
* feat: scan of images in bundle
Automation of CVE scanning requires a workflow that perform retrieval and scan of all images in the bundle.
This workflow relies on input filename specified when workflow is triggered. It usese kubeflow-ci/scripts/image/ tools to retrieve images, scan for CVEs and produce Trivy reports and scan summary file. Reports and scan summary are uploaded to the job artifacts.
Relies on the following to be merged:
canonical/seldon-core-operator#163
canonical/kubeflow-ci#80

Summary of changes:
- Workflow that performs scanning of all images in bundle.
- Added multiple steps.
- Use matrix strategy to scan multiple bundles
- Use data in artifacts creation.
- Use existing sparse checkout actions.
- Added release and risk to get images call
- Updated script name to all get all images