Update the release automation for 1.8 (#722)
* Update workflow to NOT run tests when publishing

Since currently the tests are broken we will disable running them in the
workflows. We will revert once we fix
https://github.com/canonical/oidc-gatekeeper-operator/issues/112

* Update scripts for including the 1.8 release

Update the scripts for gathering paths and test files for the 1.8
release.

* Add test files for 1.8

Copy over all files from 1.7 for 1.8. Although at the time of the commit
we will have disabled the tests, but we will introduce them nevertheless
to be there in the future.

* review: Put accepted tracks/risk as global var

* Remove redundant istio-pilot config-event triggering