ci: refactor scan-images.yaml workflow
This commit refactors the scan images workflow to use the aquasecurity/trivy-action
Github action instead of the kubeflow-ci/scripts/images/scan-images.sh. This to make
the workflow faster and more configurable based on the action's inputs.
This commit also applies a mitigation for the upstream issue aquasecurity/trivy-action#389,
as the number of DB download requests would be reduced (each image is scanned independently
in its own Github runner).